refactor: extract pure builders and crypto helper from XrplService
Move the four pure intent/operation builders to xrpl.builders.ts and
compressPublicKey to xrpl.crypto.ts, trimming xrpl.service.ts so it
focuses on orchestration. No behavioral change.
